requirements.txt
tensorflow==2.10.0
numpy
pandas
matplotlib
seaborn
plotly
pillow
scikit-learn
optree
pydot
graphviz
